Clean up Communicator interface with changes
This patch introduces naming rule to communicator API methods, as
adding suffix `_obj` for Python object transfers. Also changes
`broadcast_data` to `bcast_data` to unify broadcasting names to bcast,
which had possibility to confuse users with `broadcast_to` in
chainer.functions. Also it hides mpi_comm usage (except `Barrier()`
usage). This patch also introduce MpiCommunicatorBase to make
CommunicatorBase purely "virtual" interface class that has no
implementation, but just interface.
