+++ chainermn/communicators/hierarchical_communicator.py
@@ -1,21 +1,18 @@
 import chainer.cuda
 import math
 
-from chainermn.communicators import _base
 from chainermn.communicators import _communication_utility
 from chainermn.communicators import _memory_utility
+from chainermn.communicators import mpi_communicator_base
 from chainermn import nccl
 
 
-class HierarchicalCommunicator(_base.CommunicatorBase):
+class HierarchicalCommunicator(mpi_communicator_base.MpiCommunicatorBase):
 
     def __init__(self, mpi_comm):
         super(HierarchicalCommunicator, self).__init__(mpi_comm, use_nccl=True)
         self.gpu_buffer_a = _memory_utility.DeviceMemory()
         self.gpu_buffer_b = _memory_utility.DeviceMemory()
-
-    def broadcast_data(self, model):
-        _communication_utility.broadcast_naive(self.mpi_comm, model)
 
     def allreduce_grad(self, model):
         self._init_comms()
