+++ chainermn/communicators/naive_communicator.py
@@ -1,17 +1,13 @@
 import mpi4py.MPI
 
-from chainermn.communicators import _base
-from chainermn.communicators import _communication_utility
 from chainermn.communicators import _memory_utility
+from chainermn.communicators import mpi_communicator_base
 
 
-class NaiveCommunicator(_base.CommunicatorBase):
+class NaiveCommunicator(mpi_communicator_base.MpiCommunicatorBase):
 
     def __init__(self, mpi_comm):
         super(NaiveCommunicator, self).__init__(mpi_comm)
-
-    def broadcast_data(self, model):
-        _communication_utility.broadcast_naive(self.mpi_comm, model)
 
     def allreduce_grad(self, model):
         for param in _memory_utility.extract_params(model):
