+++ chainermn/communicators/single_node_communicator.py
@@ -1,11 +1,11 @@
 import chainer.cuda
 
-from chainermn.communicators import _base
 from chainermn.communicators import _memory_utility
+from chainermn.communicators import mpi_communicator_base
 from chainermn import nccl
 
 
-class SingleNodeCommunicator(_base.CommunicatorBase):
+class SingleNodeCommunicator(mpi_communicator_base.MpiCommunicatorBase):
 
     def __init__(self, mpi_comm):
         super(SingleNodeCommunicator, self).__init__(mpi_comm, use_nccl=True)
@@ -17,7 +17,7 @@ class SingleNodeCommunicator(_base.CommunicatorBase):
         self.gpu_buffer_a = _memory_utility.DeviceMemory()
         self.gpu_buffer_b = _memory_utility.DeviceMemory()
 
-    def broadcast_data(self, model):
+    def bcast_data(self, model):
         self._init_comms()
         stream = chainer.cuda.Stream.null
 
@@ -36,6 +36,8 @@ class SingleNodeCommunicator(_base.CommunicatorBase):
 
         _memory_utility.unpack_params(
             params, itemsize, 'data', self.gpu_buffer_a)
+
+    broadcast_data = bcast_data
 
     def allreduce_grad(self, model):
         self._init_comms()
