+++ chainermn/communicators/two_dimensional_communicator.py
@@ -2,22 +2,19 @@ import chainer.cuda
 import math
 import mpi4py.MPI
 
-from chainermn.communicators import _base
 from chainermn.communicators import _communication_utility
 from chainermn.communicators import _memory_utility
+from chainermn.communicators import mpi_communicator_base
 from chainermn import nccl
 
 
-class TwoDimensionalCommunicator(_base.CommunicatorBase):
+class TwoDimensionalCommunicator(mpi_communicator_base.MpiCommunicatorBase):
 
     def __init__(self, mpi_comm=mpi4py.MPI.COMM_WORLD):
         super(TwoDimensionalCommunicator, self).__init__(
             mpi_comm, use_nccl=True)
         self.gpu_buffer_a = _memory_utility.DeviceMemory()
         self.gpu_buffer_b = _memory_utility.DeviceMemory()
-
-    def broadcast_data(self, model):
-        _communication_utility.broadcast_naive(self.mpi_comm, model)
 
     def allreduce_grad(self, model):
         self._init_comms()
