+++ chainermn/extensions/checkpoint.py
@@ -154,7 +154,7 @@ class _MultiNodeCheckpointer(extension.Extension):
         self.files.append(filename)
 
         if len(self.files) - self.cp_interval > 5:
-            # remove older snapshots, and bcast latest list
+            # remove older snapshots, and broadcast latest list
             self._sync_file_list(remove_remainder=True)
 
     def finalize(self):
@@ -188,7 +188,7 @@ class _MultiNodeCheckpointer(extension.Extension):
         return self.stats.report()
 
     def _sync_file_list(self, remove_remainder=False):
-        file_lists = self.comm.mpi_comm.gather(self.files, root=0)
+        file_lists = self.comm.gather_obj(self.files)
 
         iters0 = None
         if self.comm.rank == 0:
@@ -211,7 +211,7 @@ class _MultiNodeCheckpointer(extension.Extension):
             else:
                 raise RuntimeError("Can't gather checkpoint file names")
 
-        iters0 = self.comm.mpi_comm.bcast(iters0, root=0)
+        iters0 = self.comm.bcast_obj(iters0)
         files = self._filenames(iters0)
 
         if remove_remainder:
