+++ examples/dcgan/train_dcgan.py
@@ -7,7 +7,6 @@ import os
 import chainer
 from chainer import training
 from chainer.training import extensions
-from mpi4py import MPI
 
 from net import Discriminator
 from net import Generator
@@ -60,9 +59,9 @@ def main():
         comm = chainermn.create_communicator('naive')
         device = -1
 
-    if comm.mpi_comm.rank == 0:
+    if comm.rank == 0:
         print('==========================================')
-        print('Num process (COMM_WORLD): {}'.format(MPI.COMM_WORLD.Get_size()))
+        print('Num process (COMM_WORLD): {}'.format(comm.size))
         if args.gpu:
             print('Using GPUs')
         print('Using {} communicator'.format(args.communicator))
