+++ examples/mnist/train_mnist_checkpoint.py
@@ -8,7 +8,6 @@ import chainer.functions as F
 import chainer.links as L
 from chainer import training
 from chainer.training import extensions
-from mpi4py import MPI
 
 import chainermn
 from chainermn.extensions import create_multi_node_checkpointer
@@ -65,9 +64,9 @@ ChainerMN example: MNIST with automatic checkpoints enabled''')
         comm = chainermn.create_communicator('naive')
         device = -1
 
-    if comm.mpi_comm.rank == 0:
+    if comm.rank == 0:
         print('==========================================')
-        print('Num process (COMM_WORLD): {}'.format(MPI.COMM_WORLD.Get_size()))
+        print('Num process (COMM_WORLD): {}'.format(comm.size))
         if args.gpu:
             print('Using GPUs')
         print('Using {} communicator'.format(args.communicator))
