+++ tests/chainermn_tests/communicator_tests/test_node_aware_communicator_base.py
@@ -5,14 +5,14 @@ import unittest
 import mpi4py.MPI
 import pytest
 
-from chainermn.communicators import _base
+from chainermn.communicators import mpi_communicator_base
 
 
-class TestCommunicatorBase(unittest.TestCase):
+class TestMpiCommunicatorBase(unittest.TestCase):
 
     def setUp(self):
         self.mpi_comm = mpi4py.MPI.COMM_WORLD
-        self.communicator = _base.CommunicatorBase(
+        self.communicator = mpi_communicator_base.MpiCommunicatorBase(
             self.mpi_comm, use_nccl=False)
 
     def test_intra_rank_with_env(self):
